refactor: component base
